requirements.txt
streamlit
streamlit_webrtc

opencv-python
mediapipe
openai-whisper
sounddevice
scipy